refactor: phase 8 — extract primitives/formatspec.rs from format.rs
